+++ RateMatrix.hpp
@@ -32,9 +32,9 @@ class RateMatrix : public DoubleMatrix {
 inline std::ostream& operator<<(std::ostream& os, const RateMatrix& m) {
 
     os << std::fixed << std::setprecision(5);
-    for (int i=0; i<m.numRows; i++)
+    for (size_t i=0; i<m.numRows; i++)
         {
-        for (int j=0; j<m.numCols; j++)
+        for (size_t j=0; j<m.numCols; j++)
             {
             if (m(i,j) > 0)
                 os << " ";
